Move ObjectData<T> to datatypes.rs file
This is done to prevent us from directly constructing it.
